Change `factorizeh*` to require square input
The LAPACK factorization routine only uses the upper left square of a
rectangular matrix anyway, so allowing a rectangular input is somewhat
confusing. If someone really wants to call `factorizeh*` on a
rectangular matrix, they can take a square slice of portion of the
rectangular matrix they want to use.

This change is necessary for `BKFactorized::det()` to be able to
return `A::Real` instead of having to check that the matrix is square
and return `Result<A::Real>`.