Add workaround for zero-size arrays with bk()
There is a bug in the LAPACKE `*sytrf` routines that makes them fail
for zero-size arrays. Basically, what happens is that when
`LAPACKE_*sytrf` queries for the optimal `lwork` parameter value,
`dsytrf.f` suggests `0`. When `LAPACKE_*sytrf` actually tries to use
`0` for `lwork`, `dsytrf.f` returns an error indicating that `lwork`
needs to be at least `1`.