Improve stride copying in generalize a bit
While still not perfect - that should be solved in upstream ndarray, the
new version is an improvement since it checks the compatibility before
attempting to construct the array.

`.into_raw_vec()` is very hard to use correctly (this is inherent, it's
an access to the raw memory model), so the new version will actually
fail to convert in some cases that were passing as silent errors before
(related to internally sliced arrays, rather uncommon).